Added New Button for Stop Debugging, Debug Working
Debugging works, needs a little work for RScrollPane to retrieve all
user-created breakpoints
Continue and Step Over is working, but not yet noticeable (through
System.out.println() only at the moment)
I also added Stop button (right next to Step Over) so that debugging can
be stopped at any time when the user feels like it.

The problem is, it focuses mainly on the very backend of the C at the
moment when you're debugging an exe file.